Revert "Batch-wise operations"
