+++ progress_bar.py
@@ -1,11 +1,12 @@
 from time import sleep
 from tqdm import tqdm
 
-def create_progress_bar(desc):
-    return tqdm(desc = desc)
+def create_progress_bars():
+    plas_bar = tqdm(desc = "plasmid      ")
+    chrom_bar = tqdm(desc = "chromosome   ")
+    ex_plas_bar = tqdm(desc = "extra plasmid")
+
+    return plas_bar, chrom_bar, ex_plas_bar
 
 def update_progress_bar(pbar,i):
     pbar.update(i)
-
-def close_progress_bar(pbar):
-    pbar.close()
